Fix the long int problem

+++ 28/main.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int find_n(int a, int b, int c)
+long find_n(long a, long b, long c)
 {
     return a * b + c % a;
 }
@@ -11,8 +11,8 @@ int main(void)
     scanf("%d", &q);
     for (int i = 0; i < q; i++)
     {
-        int a, b, c;
-        scanf("%d %d %d", &a, &b, &c);
-        printf("%d\n", find_n(a, b, c));
+        long a, b, c;
+        scanf("%ld %ld %ld", &a, &b, &c);
+        printf("%ld\n", find_n(a, b, c));
     }
 }
